backend/requirements.txt
# Direct dependencies only — transitive deps resolved by pip
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
google-genai>=1.0.0
anthropic>=0.80.0
